pyproject.toml
[tool.poetry]
name = "cash-flow"
version = "0.1.0"
description = ""
authors = ["Tamir Mandreev <tamirmandreev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
django-filter = "^25.1"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
